+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [v0.2.0] - 2021-07-17
+### Added 
+   - Add functionality to copy selected ssm value to clipboard
+   - Add functionality to copy selected ssm name to clipboard
 ## [v0.1.0] - 2021-06-15
 ### Added 
    - Basic functionality to browse SSM params and show param details